+++ src/index.js
@@ -1,8 +1,8 @@
 import './style.css';
-import { fetchPokemon} from './modules/displayPokemon';
+import { fetchPokemon } from './modules/displayPokemon';
 import pokeCounter from './modules/pokeCounter';
 
 window.onload = () => {
-    fetchPokemon();
-    pokeCounter();
-  };
+  fetchPokemon();
+  pokeCounter();
+};
