+++ src/modules/createImg.js
@@ -1,0 +1,5 @@
+export default function createPokeImage(pokeID, containerDiv) {
+  const pokeImage = document.createElement('img');
+  pokeImage.src = `https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/${pokeID}.png`;
+  containerDiv.append(pokeImage);
+}
