+++ src/style.css
@@ -88,6 +88,7 @@ a {
   bottom: 10px;
   overflow: auto;
   background-color: white;
+  display: none;
 }
 
 #close {
