+++ src/tests/pokemonCount.js
@@ -1,3 +1,3 @@
 export default function pokemonCounter(pokemonList) {
-    return pokemonList.length;
-  }
+  return pokemonList.length;
+}
